requirements.txt
numpy~=1.23.4
tensorflow~=2.10.0
Pillow~=9.3.0
matplotlib~=3.6.2